third_party/WebKit/LayoutTests/fast/css-grid-layout/grid-update-sizes-after-distributing-all-expected.txt

Check that tracks are updated after distributing the extra space for all the items in the same group span.
PASS window.getComputedStyle(gridAutoAndMaxContent, '').getPropertyValue('grid-template-columns') is "25px 65px"
PASS window.getComputedStyle(gridFixedMinContentAndFixedMaxContent, '').getPropertyValue('grid-template-columns') is "25px 85px"
PASS window.getComputedStyle(gridMinContentAndMaxContent, '').getPropertyValue('grid-template-columns') is "15px 85px"
PASS window.getComputedStyle(gridMinContentAndFixedAndMaxContent, '').getPropertyValue('grid-template-columns') is "15px 20px 105px"
PASS window.getComputedStyle(gridMaxContentAndFixedAndAuto, '').getPropertyValue('grid-template-columns') is "105px 20px 15px"
PASS window.getComputedStyle(gridRepeatTwoMaxContentAndMinContentTwoOverlapping, '').getPropertyValue('grid-template-columns') is "115px 25px 85px 15px"
PASS window.getComputedStyle(gridRepeatTwoMaxContentAndMinContentThreeOverlapping, '').getPropertyValue('grid-template-columns') is "65px 25px 115px 15px"
PASS window.getComputedStyle(gridRepeatTwoMaxContentAndMinContentTwoAndFourOverlapping, '').getPropertyValue('grid-template-columns') is "70px 30px 80px 10px"
PASS window.getComputedStyle(gridRepeatTwoAutoAndMaxContentTwoOverlapping, '').getPropertyValue('grid-template-columns') is "25px 115px 15px 85px"
PASS window.getComputedStyle(gridRepeatTwoAutoAndMaxContentThreeOverlapping, '').getPropertyValue('grid-template-columns') is "20px 115px 25px 75px"
PASS window.getComputedStyle(gridRepeatTwoAutoAndMaxContentTwoAndFourOverlapping, '').getPropertyValue('grid-template-columns') is "30px 70px 10px 80px"
PASS successfullyParsed is true

TEST COMPLETE

XXXX XXXX
XXXXX X
XXXX XXXX
XXXXX XXXXX
XX XX
XXX XX XXX
XXXX XXXX
XXXXX XXX XXXX
XXXX XXXX
XXXXX XXX XXXX
XXXX XXXX
XX XX
XXXXX XXX XXXX
XXX XX XXX
XXXXX XXX XXXX
XXX XX XXX
XXXX XXXX
XXXX XXXX
XX XX
XXXXX XXX XXXX
XXX XX XXX
X X X X X
XXXXXX
XXXX XXXX
XX XX
XXXXX XXX XXXX
XXX XX XXX
XXXXX XXX XXXX
XXX XX XXX
XXXX XXXX
XXXX XXXX
XX XX
XXXXX XXX XXXX
XXX XX XXX
X X X X X
XXXXXX
